Pricing made usable, car type added, discount percentage changed to double type

+++ src/main/java/com/example/demo/customers/Customer.java
@@ -25,5 +25,5 @@ public class Customer {
     private String phoneNumber;
     @NotBlank(message = "first name is mandatory")
     @Column(length = 3) // 0-100
-    private Short discountPercentage = 0;
+    private double discountPercentage = 0;
 }
